[Feat] #67 - 미션 추가, 아이템 UI 보이게하기, 블루베리 추가, X 로직 추가, npc난이도 조절 추가, 미션 클리어 후 리프레시 추가

+++ Source/Module/NPC/NPC.h
@@ -9,7 +9,7 @@ void npcPutDownCard(Game *game);
 void npcPutDownCardNotAdvantage(Game *game);
 void npcPutDownRightCardAdvantage(game);
 void npcPutDownLeftCardAdvantage(Game *game);
-double getTimeWhenNPCDrawCard(void);
+int getTimeWhenNPCDrawCard(void);
 // npc 난이도 조절
 void raiseDifficultyNPCAdvantage(NPC *npc);
 void returnRaiseDifficultyNPCAdvantage(NPC *npc);
